feat: Map to show clinics, Test map for functionality #2
Have updated the map to show markers for each clinic and have a popup which leads to their specific appointment pages.
Currently the page has a function with mock data in it which shall be deleten after API integration. Have added a router which will handle url management for the web app.
Changed the Vite config so ports are not automatically exposed for the rest of the network which may pose security risks.
Have added a home page to show the map, may add a separate map page later in case a home page is required.